pyproject.toml
[tool.poetry]
name = "autor-mentoria-marcos"
version = "0.1.0"
description = ""
authors = ["victorhausen <hausen.victor@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.6"
Django = "^3.1.6"
django-restframework = "^0.0.1"
djangorestframework = "^3.12.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
